backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.7
pandas>=2.1.0
numpy>=1.25.0
pydantic>=2.5.0
python-dotenv>=1.0.0
alembic>=1.12.0
requests>=2.31.0
python-multipart>=0.0.6
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
factory-boy>=3.3.0 